+++ resources-ktx/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **Context**: added type checks to extensions `resolveColor`
 - **Context**: new extensions `resolveDimension`, `resolveDimensionPixelSize` and `resolveDimensionPixelOffset`
+- **Context**: new extension `resolveBoolean`
 
 ### Dependencies